+++ iceoryx_hoofs/include/iceoryx_hoofs/cxx/helplets.hpp
@@ -356,7 +356,7 @@ constexpr T into(const F value) noexcept;
     }                                                                                                                  \
                                                                                                                        \
   private:                                                                                                             \
-    type m_##name = defaultValue;
+    type m_##name{defaultValue};
 
 } // namespace cxx
 } // namespace iox