+++ iceoryx_hoofs/include/iceoryx_hoofs/internal/posix_wrapper/shared_memory_object/shared_memory.hpp
@@ -1,5 +1,5 @@
 // Copyright (c) 2019 by Robert Bosch GmbH. All rights reserved.
-// Copyright (c) 2021 by Apex.AI Inc. All rights reserved.
+// Copyright (c) 2021 - 2022 by Apex.AI Inc. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,10 +17,11 @@
 #ifndef IOX_HOOFS_POSIX_WRAPPER_SHARED_MEMORY_OBJECT_SHARED_MEMORY_HPP
 #define IOX_HOOFS_POSIX_WRAPPER_SHARED_MEMORY_OBJECT_SHARED_MEMORY_HPP
 
+#include "iceoryx_hoofs/cxx/expected.hpp"
+#include "iceoryx_hoofs/cxx/filesystem.hpp"
+#include "iceoryx_hoofs/cxx/helplets.hpp"
 #include "iceoryx_hoofs/cxx/optional.hpp"
 #include "iceoryx_hoofs/cxx/string.hpp"
-#include "iceoryx_hoofs/design_pattern/creation.hpp"
-#include "iceoryx_hoofs/platform/mman.hpp"
 
 #include <cstdint>
 
@@ -54,7 +55,7 @@ static constexpr const char* OPEN_MODE_STRING[] = {
 enum class SharedMemoryError
 {
     EMPTY_NAME,
-    NAME_WITHOUT_LEADING_SLASH,
+    INVALID_FILE_NAME,
     INSUFFICIENT_PERMISSIONS,
     DOES_EXIST,
     PROCESS_LIMIT_OF_OPEN_FILES_REACHED,
@@ -74,7 +75,7 @@ enum class SharedMemoryError
 ///        shm_open, shm_unlink etc.
 ///        It must be used in combination with MemoryMap (or manual mmap calls)
 //         to gain access to the created/opened shared memory
-class SharedMemory : public DesignPattern::Creation<SharedMemory, SharedMemoryError>
+class SharedMemory
 {
   public:
     static constexpr uint64_t NAME_SIZE = platform::IOX_MAX_SHM_NAME_LENGTH;
@@ -103,28 +104,15 @@ class SharedMemory : public DesignPattern::Creation<SharedMemory, SharedMemoryEr
     ///         SharedMemoryError when the underlying shm_unlink call failed.
     static cxx::expected<bool, SharedMemoryError> unlinkIfExist(const Name_t& name) noexcept;
 
-    friend class DesignPattern::Creation<SharedMemory, SharedMemoryError>;
+    friend class SharedMemoryBuilder;
 
   private:
-    /// @brief constructs or opens existing shared memory
-    /// @param[in] name the name of the shared memory, must start with a leading /
-    /// @param[in] accessMode defines if the shared memory is mapped read only or with read write rights
-    /// @param[in] openMode states how the shared memory is created/opened
-    /// @param[in] permissions the permissions the shared memory should have
-    /// @param[in] size the size in bytes of the shared memory
-    SharedMemory(const Name_t& name,
-                 const AccessMode accessMode,
-                 const OpenMode openMode,
-                 const mode_t permissions,
-                 const uint64_t size) noexcept;
+    SharedMemory(const Name_t& name, const int handle, const bool hasOwnership) noexcept;
 
-    bool
-    open(const AccessMode accessMode, const OpenMode openMode, const mode_t permissions, const uint64_t size) noexcept;
     bool unlink() noexcept;
     bool close() noexcept;
     void destroy() noexcept;
     void reset() noexcept;
-    static int getOflagsFor(const AccessMode accessMode, const OpenMode openMode) noexcept;
 
     static SharedMemoryError errnoToEnum(const int32_t errnum) noexcept;
 
@@ -132,6 +120,34 @@ class SharedMemory : public DesignPattern::Creation<SharedMemory, SharedMemoryEr
     int m_handle{INVALID_HANDLE};
     bool m_hasOwnership{false};
 };
+
+class SharedMemoryBuilder
+{
+    /// @brief A valid file name for the shared memory with the restriction that
+    ///        no leading dot is allowed since it is not compatible with every
+    ///        file system
+    IOX_BUILDER_PARAMETER(SharedMemory::Name_t, name, "")
+
+    /// @brief Defines if the memory should be mapped read only or with write access.
+    ///        A read only memory section will cause a segmentation fault when written to.
+    IOX_BUILDER_PARAMETER(AccessMode, accessMode, AccessMode::READ_ONLY)
+
+    /// @brief Defines how the shared memory is acquired
+    IOX_BUILDER_PARAMETER(OpenMode, openMode, OpenMode::OPEN_EXISTING)
+
+    /// @brief Defines the access permissions of the shared memory
+    IOX_BUILDER_PARAMETER(cxx::perms, filePermissions, cxx::perms::none)
+
+    /// @brief Defines the size of the shared memory
+    IOX_BUILDER_PARAMETER(uint64_t, size, 0U)
+
+  public:
+    /// @brief creates a valid SharedMemory object. If the construction failed the expected
+    ///        contains an enum value describing the error.
+    /// @return expected containing SharedMemory on success otherwise SharedMemoryError
+    cxx::expected<SharedMemory, SharedMemoryError> create() noexcept;
+};
+
 } // namespace posix
 } // namespace iox